backend/requirements.txt
django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
langchain-openai>=0.3.1
langchain-core>=0.3.1

